Export runs that don't generate files can now also show that they have run and whether they are up to date.
It used to be the case that all export runs generated a file, and the
check to see if they had run successfully some time in the past looked
for the presence of this file. But the 'Export to Lizard' export run
exports data to elsewhere, and doesn't have a file. There is now a new
property, 'available', and for runs without generated files it merely
checks for a created_at date. User interface now uses that, except for
the Download link, which still needs a 'present' file.